// src/app/api/docstorageApi.client.ts
export interface DocumentByIdResponseDto {
    documentId?: string | undefined;
    documentCode?: string | undefined;
    mainIndex?: string | undefined;
    searchIndex?: string | undefined;
    documentTypeName?: string | undefined;
    documentSubTypeName?: string | undefined;
    fileCount?: number | undefined;
    documentFiles?: DocumentFileResponseDto[] | undefined;
    documentIndexs?: DocumentIndexValueByDocumentIdResponseDto[] | undefined;
}
// ...
export interface DocumentByIdResponseDtoServiceResponse {
    data?: DocumentByIdResponseDto;
    isSuccess?: boolean;
    message?: string | undefined;
    code?: number | undefined;
    exceptionMessage?: any | undefined;
    serverDateTime?: dayjs.Dayjs;
    totalAmountRecords?: number | undefined;
    totalAmountPages?: number | undefined;
    currentPage?: number | undefined;
    recordsPerPage?: number | undefined;
    pageIndex?: number | undefined;
}
// ...
export interface DocumentFileResponseDto {
    documentFileId?: string | undefined;
    s3IsUploaded?: boolean | undefined;
    s3Bucket?: string | undefined;
    s3Key?: string | undefined;
    physicalPath?: string | undefined;
    fileName?: string | undefined;
    fileURL?: string | undefined;
    pathThumbnailImg?: string | undefined;
    pathFullDoc?: string | undefined;
    isImage?: boolean;
    createdDate?: dayjs.Dayjs | undefined;
    createdByUserId?: number | undefined;
    createByUserCode?: string | undefined;
    createdByUserName?: string | undefined;
    deletedByUserId?: number | undefined;
    deletedByUserCode?: string | undefined;
    deletedByUserName?: string | undefined;
    deletedDate?: dayjs.Dayjs | undefined;
    isActive?: boolean | undefined;
    recoveredDate?: dayjs.Dayjs | undefined;
    recoveredByUserId?: number | undefined;
    recoveredByUserCode?: string | undefined;
    recoveredByUserName?: string | undefined;
    updatedDate?: dayjs.Dayjs | undefined;
}
// ...
export interface DocumentIndexValueByDocumentIdResponseDto {
    documentIndexName?: string | undefined;
    value?: string | undefined;
}
// ...
export interface UploadDocumentFilesResponseDto {
    documentFileId?: string;
    fileName?: string | undefined;
}

export interface UploadResponseDto {
    documentId?: string;
    files?: UploadDocumentFilesResponseDto[] | undefined;
}

export interface UploadResponseDtoServiceResponse {
    data?: UploadResponseDto;
    isSuccess?: boolean;
    message?: string | undefined;
    code?: number | undefined;
    exceptionMessage?: any | undefined;
    serverDateTime?: dayjs.Dayjs;
    totalAmountRecords?: number | undefined;
    totalAmountPages?: number | undefined;
    currentPage?: number | undefined;
    recordsPerPage?: number | undefined;
    pageIndex?: number | undefined;
}

export interface FileParameter {
    data: any;
    fileName: string;
}